organize View and Controller of SubPlayer and MainPlayer
