--- src/main/java/com/ouroboros/sleepingqueen/MainPlayerCardField.java
@@ -1,6 +1,0 @@
-package com.ouroboros.sleepingqueen;
-
-
-public class MainPlayerCardField {
-
-}
