+++ src/main/java/com/ouroboros/sleepingqueen/mainPlayer/BoardMenuController.java
@@ -1,4 +1,4 @@
-package com.ouroboros.sleepingqueen;
+package com.ouroboros.sleepingqueen.mainPlayer;
 
 public class BoardMenuController {
     // TODO: Implement this class
