+++ src/main/java/com/ouroboros/sleepingqueen/mainPlayer/MainPlayerCardField.java
@@ -1,0 +1,6 @@
+package com.ouroboros.sleepingqueen.mainPlayer;
+
+
+public class MainPlayerCardField {
+
+}
