+++ src/main/java/com/ouroboros/sleepingqueen/mainPlayer/MainPlayerQueenField.java
@@ -1,4 +1,4 @@
-package com.ouroboros.sleepingqueen;
+package com.ouroboros.sleepingqueen.mainPlayer;
 
 public class MainPlayerQueenField {
     // TODO: Implement this class
