+++ src/main/java/module-info.java
@@ -20,4 +20,6 @@ module com.ouroboros.sleepingqueen {
     opens com.ouroboros.sleepingqueen.testScreen to javafx.fxml;
     exports com.ouroboros.sleepingqueen.subPlayer;
     opens com.ouroboros.sleepingqueen.subPlayer to javafx.fxml;
+    exports com.ouroboros.sleepingqueen.mainPlayer;
+    opens com.ouroboros.sleepingqueen.mainPlayer to javafx.fxml;
 }
