+++ src/main/resources/com/ouroboros/sleepingqueen/view/board-view.fxml
@@ -20,27 +20,27 @@
         <HBox layoutX="226.0" layoutY="7.0" prefHeight="562.0" prefWidth="828.0" AnchorPane.topAnchor="0.0">
             <children>
                 <HBox prefHeight="559.0" prefWidth="323.0" />
-                <fx:include fx:id="queenField" source="queen-field.fxml" />
+                <fx:include fx:id="queenField" source="boardView/queen-field.fxml" />
             </children>
         </HBox>
         <HBox prefHeight="559.0" prefWidth="323.0" AnchorPane.leftAnchor="477.0" AnchorPane.rightAnchor="477.0" AnchorPane.topAnchor="0.0">
             <children>
-                <fx:include fx:id="deckField" source="deck-on-board.fxml" />
+                <fx:include fx:id="deckField" source="boardView/deck-on-board.fxml" />
             </children>
         </HBox>
       <HBox layoutY="564.0" prefHeight="153.0" prefWidth="291.0">
          <children>
-             <fx:include fx:id="mainPlayerQueenField" source="main-player-queen-field.fxml" />
+             <fx:include fx:id="mainPlayerQueenField" source="mainPlayerView/main-player-queen-field.fxml" />
          </children>
       </HBox>
       <HBox layoutX="294.0" layoutY="564.0" prefHeight="158.0" prefWidth="743.0">
          <children>
-            <fx:include fx:id="mainPlayerCardField" source="main-player-card-field.fxml" />
+            <fx:include fx:id="mainPlayerCardField" source="mainPlayerView/main-player-card-field.fxml" />
          </children>
       </HBox>
       <HBox layoutX="1040.0" layoutY="564.0" prefHeight="158.0" prefWidth="240.0">
          <children>
-            <fx:include fx:id="Menu" source="main-player-name-and-menu.fxml" />
+            <fx:include fx:id="Menu" source="mainPlayerView/main-player-name-and-menu.fxml" />
          </children>
       </HBox>
     </children>
