+++ src/main/resources/com/ouroboros/sleepingqueen/view/boardView/queen-field.fxml
