+++ src/main/resources/com/ouroboros/sleepingqueen/view/mainPlayerView/main-player-queen-card.fxml
@@ -11,7 +11,7 @@
     <children>
         <ImageView id="main-player-queen-card" fitHeight="102.0" fitWidth="62.0" pickOnBounds="true">
             <image>
-                <Image url="@../cardImg/queen-09.jpg" />
+                <Image url="@../../cardImg/queen-09.jpg" />
             </image>
         </ImageView>
     </children>
